# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Color Gopher API",
    "description": "A color converter and palette generator"
  },
  "host": "localhost:5500",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {},
  "consumes": [
    "application / json"
  ],
  "produces": [
    "application / json"
  ],
  "paths": {
...
    "/gopher/rgb-to-all": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "r",
            "in": "query",
            "type": "string"
          },
          {
            "name": "g",
            "in": "query",
            "type": "string"
          },
          {
            "name": "b",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/gopher/hsv-to-all": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "h",
            "in": "query",
            "type": "string"
          },
          {
            "name": "s",
            "in": "query",
            "type": "string"
          },
          {
            "name": "v",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/gopher/hsl-to-all": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "h",
            "in": "query",
            "type": "string"
          },
          {
            "name": "s",
            "in": "query",
            "type": "string"
          },
          {
            "name": "l",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/gopher/cmyk-to-all": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "c",
            "in": "query",
            "type": "string"
          },
          {
            "name": "m",
            "in": "query",
            "type": "string"
          },
          {
            "name": "y",
            "in": "query",
            "type": "string"
          },
          {
            "name": "k",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/gopher/hex-to-all": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "hex",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/gopher/name-to-all": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/gopher/palette-complement": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "r",
            "in": "query",
            "type": "string"
          },
          {
            "name": "g",
            "in": "query",
            "type": "string"
          },
          {
            "name": "b",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/gopher/palette-split-complement": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "r",
            "in": "query",
            "type": "string"
          },
          {
            "name": "g",
            "in": "query",
            "type": "string"
          },
          {
            "name": "b",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/gopher/palette-adjacent": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "r",
            "in": "query",
            "type": "string"
          },
          {
            "name": "g",
            "in": "query",
            "type": "string"
          },
          {
            "name": "b",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/gopher/palette-analagous": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "r",
            "in": "query",
            "type": "string"
          },
          {
            "name": "g",
            "in": "query",
            "type": "string"
          },
          {
            "name": "b",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/gopher/palette-phi-triad": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "r",
            "in": "query",
            "type": "string"
          },
          {
            "name": "g",
            "in": "query",
            "type": "string"
          },
          {
            "name": "b",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/gopher/palette-equi-triad": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "r",
            "in": "query",
            "type": "string"
          },
          {
            "name": "g",
            "in": "query",
            "type": "string"
          },
          {
            "name": "b",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/gopher/palette-phi-tetrad": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "r",
            "in": "query",
            "type": "string"
          },
          {
            "name": "g",
            "in": "query",
            "type": "string"
          },
          {
            "name": "b",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/gopher/palette-equi-tetrad": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "r",
            "in": "query",
            "type": "string"
          },
          {
            "name": "g",
            "in": "query",
            "type": "string"
          },
          {
            "name": "b",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/gopher/palette-mono-scale": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "r",
            "in": "query",
            "type": "string"
          },
          {
            "name": "g",
            "in": "query",
            "type": "string"
          },
          {
            "name": "b",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/gopher/palette-interpolate": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "r1",
            "in": "query",
            "type": "string"
          },
          {
            "name": "g1",
            "in": "query",
            "type": "string"
          },
          {
            "name": "b1",
            "in": "query",
            "type": "string"
          },
          {
            "name": "r2",
            "in": "query",
            "type": "string"
          },
          {
            "name": "g2",
            "in": "query",
            "type": "string"
          },
          {
            "name": "b2",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/gopher/rgb-to-hsv": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "r",
            "in": "query",
            "type": "string"
          },
          {
            "name": "g",
            "in": "query",
            "type": "string"
          },
          {
            "name": "b",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/gopher/hsv-to-rgb": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "h",
            "in": "query",
            "type": "string"
          },
          {
            "name": "s",
            "in": "query",
            "type": "string"
          },
          {
            "name": "v",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/gopher/rgb-to-cmyk": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "r",
            "in": "query",
            "type": "string"
          },
          {
            "name": "g",
            "in": "query",
            "type": "string"
          },
          {
            "name": "b",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/gopher/cmyk-to-rgb": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "c",
            "in": "query",
            "type": "string"
          },
          {
            "name": "m",
            "in": "query",
            "type": "string"
          },
          {
            "name": "y",
            "in": "query",
            "type": "string"
          },
          {
            "name": "k",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/gopher/rgb-to-hsl": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "r",
            "in": "query",
            "type": "string"
          },
          {
            "name": "g",
            "in": "query",
            "type": "string"
          },
          {
            "name": "b",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/gopher/hsl-to-rgb": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "h",
            "in": "query",
            "type": "string"
          },
          {
            "name": "s",
            "in": "query",
            "type": "string"
          },
          {
            "name": "l",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/gopher/rgb-to-hex": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "r",
            "in": "query",
            "type": "string"
          },
          {
            "name": "g",
            "in": "query",
            "type": "string"
          },
          {
            "name": "b",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/gopher/hex-to-rgb": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "hex",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/gopher/hsv-to-hsl": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "h",
            "in": "query",
            "type": "string"
          },
          {
            "name": "s",
            "in": "query",
            "type": "string"
          },
          {
            "name": "v",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/gopher/hsl-to-hsv": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "h",
            "in": "query",
            "type": "string"
          },
          {
            "name": "s",
            "in": "query",
            "type": "string"
          },
          {
            "name": "l",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/gopher/hsl-to-cmyk": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "h",
            "in": "query",
            "type": "string"
          },
          {
            "name": "s",
            "in": "query",
            "type": "string"
          },
          {
            "name": "l",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/gopher/cmyk-to-hsl": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "c",
            "in": "query",
            "type": "string"
          },
          {
            "name": "m",
            "in": "query",
            "type": "string"
          },
          {
            "name": "y",
            "in": "query",
            "type": "string"
          },
          {
            "name": "k",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/gopher/hsv-to-cmyk": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "h",
            "in": "query",
            "type": "string"
          },
          {
            "name": "s",
            "in": "query",
            "type": "string"
          },
          {
            "name": "v",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/gopher/hex-to-hsv": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "hex",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/gopher/hsl-to-hex": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "h",
            "in": "query",
            "type": "string"
          },
          {
            "name": "s",
            "in": "query",
            "type": "string"
          },
          {
            "name": "l",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/gopher/hex-to-hsl": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "hex",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/gopher/hex-to-cmyk": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "hex",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}